# sigverify using TransactionMeta