--- a/perf/src/lib.rs
+++ b/perf/src/lib.rs
@@ -18,7 +18,6 @@ extern crate lazy_static;
 extern crate log;
 
 #[cfg(test)]
-#[macro_use]
 extern crate assert_matches;
 
 #[macro_use]