#### Problem
- Can re-use this parsing code and do some additional *simple* checks before signature verification
- This let's us reduce code duplication and also not waste time on malformed packets

#### Summary of Changes
- perf/sigverify using `TransactionMeta`

Fixes #
<!-- OPTIONAL: Feature Gate Issue: # -->
<!-- Don't forget to add the "feature-gate" label -->