@apfitzge 2024-08-14
Most cases are around the same, but the `low_packets_large_batch` seems to have worse performance.

<details>
<summary>Benchmark Results</summary>

```
master:
test bench_get_offsets                          ... bench:      20,261 ns/iter (+/- 405)
test bench_sigverify_high_packets_large_batch   ... bench:  15,607,166 ns/iter (+/- 882,704)
test bench_sigverify_high_packets_small_batch   ... bench:  15,958,404 ns/iter (+/- 941,548)
test bench_sigverify_low_packets_large_batch    ... bench:   3,593,510 ns/iter (+/- 58,357)
test bench_sigverify_low_packets_small_batch    ... bench:   3,583,400 ns/iter (+/- 56,680)
test bench_sigverify_medium_packets_large_batch ... bench:   4,141,554 ns/iter (+/- 639,084)
test bench_sigverify_medium_packets_small_batch ... bench:   4,230,812 ns/iter (+/- 756,122)
test bench_sigverify_simple                     ... bench:   3,678,835 ns/iter (+/- 45,763)
test bench_sigverify_uneven                     ... bench:   1,824,205 ns/iter (+/- 73,714)

PR:
test bench_get_offsets                          ... bench:      20,990 ns/iter (+/- 118)
test bench_sigverify_high_packets_large_batch   ... bench:  15,569,566 ns/iter (+/- 1,360,445)
test bench_sigverify_high_packets_small_batch   ... bench:  15,804,020 ns/iter (+/- 1,174,673)
test bench_sigverify_low_packets_large_batch    ... bench:   4,122,681 ns/iter (+/- 1,064,313)
test bench_sigverify_low_packets_small_batch    ... bench:   3,705,250 ns/iter (+/- 62,343)
test bench_sigverify_medium_packets_large_batch ... bench:   4,078,141 ns/iter (+/- 688,364)
test bench_sigverify_medium_packets_small_batch ... bench:   4,032,270 ns/iter (+/- 590,441)
test bench_sigverify_simple                     ... bench:   3,748,470 ns/iter (+/- 46,503)
test bench_sigverify_uneven                     ... bench:   1,812,433 ns/iter (+/- 49,907)
```

</details>